Extend LaTeX functionality to include math text
In LaTeX, math-mode text is marked up as surrounded by escaped
parentheses: \(x = 1\). This is an alternative to the raw-TeX dollar
sign notation: $x = 1$.

The standard functionality of vim-surround doesn't lend itself to
quickly enclosing text with this two-character pattern. This commit
extends the built-in handling of LaTeX \begin{env} ... \end{env}
enclosures to do this when 'env' is not specified at the prompt.

To use it, select the text you want to surround, hit `S` and `l` (or
`\`) to enter LaTeX mode. When prompted for an environment name
(`\begin{`) just hit enter. The escaped parentheses are used instead.

This commit was originally authored on December 19, 2015.